% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-response-url.R
\name{add_response_url_from}
\alias{add_response_url_from}
\title{Add a \code{reapr_doc} response prefix URL to a data frame}
\usage{
add_response_url_from(xdf, x)
}
\arguments{
\item{xdf}{a data frame}

\item{x}{a \code{reapr_doc}}
}
\description{
Add a \code{reapr_doc} response prefix URL to a data frame
}
\examples{
x <- reap_url("http://books.toscrape.com/")

# good ol' R
add_response_url_from(
  as.data.frame(x$tag$a),
  x
)

\dontrun{
# piping
as_tibble(x$tag$a) \%>\%
  add_response_url_from(x)
}
}
