% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mill.R
\name{mill}
\alias{mill}
\title{Turn a \code{reapr_doc} into plain text without cruft}
\usage{
mill(x)
}
\arguments{
\item{x}{a \code{reapr_doc}}
}
\value{
a character vector of plain text with no HTML
}
\description{
Plain text extraction is accomplished via the following idiom: first,
an attempt is made to use an XSLT style sheet to select only the best
target nodes for extraction. On some malformed HTML content this
results in an empty document. When that occurs a less conservative
approach is taken with a simple XPath that is desgined to capture
all \code{<body>} elements that are not \code{<script>} tags. This is imperfect
but does provide fairly decent results when the preferred method fails.
}
