% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-url.R
\name{print.reapr_doc}
\alias{print.reapr_doc}
\title{Print for \code{reapr_doc} objects}
\usage{
\method{print}{reapr_doc}(x, ...)
}
\arguments{
\item{x}{a \code{reapr_doc}}

\item{...}{unused}
}
\description{
Print for \code{reapr_doc} objects
}
\keyword{internal}
