% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reapr-find.R
\name{reap_find}
\alias{reap_find}
\title{Return node(s) based on selector}
\usage{
reap_find(doc, css, xpath)
}
\arguments{
\item{doc}{a \code{reapr_html_xml2} document}

\item{css, xpath}{specify either (i.e. not both)}
}
\description{
Return node(s) based on selector
}
