% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-html.R
\name{reap_html}
\alias{reap_html}
\title{Read HTML from a URL}
\usage{
reap_html(url, as = c("parsed", "html"), encoding = "", ...)
}
\arguments{
\item{url}{a URL to retrieve via \code{HTTP} \code{GET}}

\item{as}{one of \code{parsed} which will return an \code{xml2::html_document} or \code{html}
which will return a character vector of un-\code{libxml2}-mangled HTML.}

\item{encoding}{override any passed-in (or server-neglected) encoding. Default
is \code{""} to let use the default behaviour of \code{\link[httr:GET]{httr::GET()}}.}

\item{...}{other parameters passed on to \code{\link[httr:GET]{httr::GET()}}}
}
\description{
Alternative to \code{\link[rvest:read_html]{rvest::read_html()}}. It uses \code{\link[httr:GET]{httr::GET()}} under the hood
so you can more easily and consistently specify browser-like configurations
to an HTTP resource.
}
