% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reap-nodes.R
\name{reap_nodes}
\alias{reap_nodes}
\alias{reap_node}
\title{Reap nodes from an reaped HTML document}
\usage{
reap_nodes(x, xpath, css)

reap_node(x, xpath, css)
}
\arguments{
\item{x}{A \code{reapr_doc} or anything \code{rvest::html_nodes()} takes.}

\item{xpath, css}{either an XPath query (string) or CSS selector; NOTE the
order difference.}
}
\description{
Provides simialar functionality to \code{rvest::html_nodes()} except when
a \code{reapr_doc} is passed in where it will then test for the validity of
the pre-parsed HTML content and regenerate the parse tree if the pointer
is invalid. Another major difference is that it prefers XPath queries over
CSS selectors so the \code{xpath} and \code{css} named (yet positional) parameters
are in a different order than in their \code{rvest} cousins.
}
\examples{
x <- reap_url("http://r-project.org/")
reap_text(x$tag$div)
reap_nodes(x, ".//*") \%>\% reap_name()
x$tag$div \%>\% reap_children()
reap_attrs(x$tag$div)
}
