% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reapr-head.R
\name{reap_title}
\alias{reap_title}
\title{Return the \code{<title>} tag value}
\usage{
reap_title(doc)
}
\arguments{
\item{doc}{a \code{reapr_html_xml2} document}
}
\description{
Return the \code{<title>} tag value
}
