% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reapr-package.R
\docType{package}
\name{reapr}
\alias{reapr}
\alias{reapr-package}
\title{Reap Information from Websites}
\description{
This will eventually be a clever description about web scraping with reapr.
}
\details{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/reapr}
\item BugReports: \url{https://gitlab.com/hrbrmstr/reapr/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
